fix: Configure azure auth plugin
If we use azure auth plugin, we have to make sure it's configured
correctly, at the same time calling this method for all plugins may be
too much. I could introduce a new field like `plugin_type` but I figured
checking the config for resource url seems like a good test to figure
out if this is a azure auth plugin.

Docs:
* https://developer.hashicorp.com/vault/docs/auth/azure#configuration
* https://github.com/hashicorp/vault-plugin-auth-azure/blob/247edaf54a3e3f5a3e57f5e27ff154d41ba16368/bootstrap/terraform/function/configure.sh#L30

Signed-off-by: Michał Jankowski <michal.jankowski@datadoghq.com>